refactor: break JetStream client dependency cycle
