+++ natskell.cabal
@@ -76,7 +76,7 @@ library jetstream
 library jetstream-internal
   import:         shared
   build-depends:
-    natskell,
+    natskell-internal,
     aeson >= 2.1 && < 2.3,
     base64-bytestring >= 1.2 && < 1.3,
     bytestring >= 0.10 && < 0.13,
@@ -128,6 +128,7 @@ library natskell-internal
     text (>= 1.2 && < 1.3) || (>= 2.0 && < 2.2)
 
   exposed-modules:
+    Client.API
     Auth.Types
     Auth.None
     Auth.Token
